requirements.txt
requests==2.22.0
google-auth==1.6.3
google-auth-oauthlib==0.4.0
